backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
PyJWT==2.9.0
Pillow==10.4.0
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1